PmagPy Cookbook files as of May 24, 2013
These are all the files associated with the most recent update to the
PmagPy cookbook as written and compiled by Lisa Tauxe. Note that the
PmagPy cookbook is documentation updated from that in the book entitled
Essentials of Paleomagnetism by Tauxe et al., (2010). This cookbook was
designed as a companion website to the book Essentials of
Paleomagnetism, 2nd Web Edition. Chapter references to this companion
book are, for example, “Essentials Chapter X”.